Makefile: Add fmt-check and fmt-fix targets
`make fmt-check` will print out the names of source files that need to
be reformatted and exit with non-zero exitcode in case any file is not
correctly formatted.

`make fmt-fix` runs `gofmt -w -s .` to fix source code formatting.

The main Makefile no runs test, fmt-check, and fmt-fix on the entire
repo instead of folder-by-folder.  When running test or fmt-check it'll
give all the errors instead of just the first folder that fails.