prometheus skydive connector
This is a sample exporter to expose skydive monitoring information to prometheus.

For each captured flow, package the number of bytes transferred
into a structure that can be consumed by prometheus.
This is performed per connection that is captured.
Users may make changes to the code to export additional features.

Cache the entries and delete from the report those conenctions that have had no recent traffic.